@@ -3,18 +3,16 @@ package com.swiftgateicthub.evote;
 import android.app.AlertDialog;
 import android.content.DialogInterface;
 import android.content.Intent;
-import android.support.annotation.NonNull;
-import android.support.annotation.Nullable;
-import android.support.design.widget.Snackbar;
-import android.support.v7.app.AppCompatActivity;
+import androidx.annotation.NonNull;
+
+import com.google.android.material.snackbar.Snackbar;
+import androidx.appcompat.app.AppCompatActivity;
 import android.os.Bundle;
 import android.text.TextUtils;
-import android.util.Log;
 import android.view.View;
 import android.widget.Button;
 import android.widget.ProgressBar;
 import android.widget.TextView;
-import android.widget.Toast;
 
 import com.google.android.gms.tasks.OnCompleteListener;
 import com.google.android.gms.tasks.Task;
@@ -110,8 +108,7 @@ public class SignupActivity extends AppCompatActivity {
         progressBar.setVisibility(View.GONE);
     }
 
-    private void sendVerificationEmail()
-    {
+    private void sendVerificationEmail() {
         FirebaseUser user = FirebaseAuth.getInstance().getCurrentUser();
 
         user.sendEmailVerification()
@@ -133,19 +130,16 @@ public class SignupActivity extends AppCompatActivity {
                             });
                             verifyAcct.show();
                             // after email is sent just logout the user and finish this activity
-                            FirebaseAuth.getInstance().signOut();
+//                            FirebaseAuth.getInstance().signOut();
                             startActivity(new Intent(SignupActivity.this, LoginActivity.class));
                             finish();
-                        }
-                        else
-                        {
+                        } else {
                             // email not sent, so display message and restart the activity or do whatever you wish to do
                             //restart this activity
                             overridePendingTransition(0, 0);
                             finish();
                             overridePendingTransition(0, 0);
                             startActivity(getIntent());
-
                         }
                     }
                 });
